Initial debian script
This patch adds support for the first debianization of the vecpf package.

With this debianization scripts, two binary packages are generated. One, called
libvecpf1 that contains the binaries, and one for development files.

This commit came from the github repository leitao/libvecpf_deb.git and the
keys contributors were myself and Frederic Bonnard.